backend/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
openai
numpy
scipy
silero
silero_stress
elevenlabs
python-jose 
passlib[argon2]
phonemizer
epitran
torchaudio
packaging
rapidfuzz
python-multipart
mutagen
